Ensure midonet_gateway works properly
midonet_gateway will not work properly if the name of the resource
is not equivalent with the actual hostname. This makes it impossible
to call midonet_gateway more than once on a host. We need to call
the function more than once on a host because it doesn't allow to
talk to different BGP peers from different source addresses. Until
puppet-midonet is made more flexible by upstream, we need to work
around this by weighting the test formalia a little bit.